remove pops file
