+++ heat/core/__init__.py
@@ -21,7 +21,6 @@ from . import random
 from .relational import *
 from .rounding import *
 from .sanitation import *
-from ._pops import *
 from .statistics import *
 from .stride_tricks import *
 from .dndarray import *
