+++ heat/core/manipulations.py
@@ -23,7 +23,7 @@ from . import tiling
 from . import types
 from . import _operations
 
-from ._pops import POps
+from ._operations import POps
 
 __all__ = [
     "balance",
@@ -4427,3 +4427,50 @@ def topk(
         out[1]._DNDarray__dtype = types.int64
 
     return final_array, final_indices
+
+
+if HAVE_MPI:
+
+    def mpi_topk(a, b, mpi_type):
+        """
+        MPI function for distributed :func:`topk`
+        """
+        # Parse Buffer
+        a_parsed = torch.from_numpy(np.frombuffer(a, dtype=np.float64))
+        b_parsed = torch.from_numpy(np.frombuffer(b, dtype=np.float64))
+
+        # Collect metadata from Buffer
+        k = int(a_parsed[0].item())
+        dim = int(a_parsed[1].item())
+        largest = bool(a_parsed[2].item())
+        sorted = bool(a_parsed[3].item())
+
+        # Offset is the length of the shape on the buffer
+        len_shape_a = int(a_parsed[4])
+        shape_a = a_parsed[5 : 5 + len_shape_a].int().tolist()
+        len_shape_b = int(b_parsed[4])
+        shape_b = b_parsed[5 : 5 + len_shape_b].int().tolist()
+
+        # separate the data into values, indices
+        a_values, a_indices = a_parsed[len_shape_a + 5 :].chunk(2)
+        b_values, b_indices = b_parsed[len_shape_b + 5 :].chunk(2)
+
+        # reconstruct the flattened data by shape
+        a_values = a_values.reshape(shape_a)
+        a_indices = a_indices.reshape(shape_a)
+        b_values = b_values.reshape(shape_b)
+        b_indices = b_indices.reshape(shape_b)
+
+        # concatenate the data to actually run topk on
+        values = torch.cat((a_values, b_values), dim=dim)
+        indices = torch.cat((a_indices, b_indices), dim=dim)
+
+        result, k_indices = torch.topk(values, k, dim=dim, largest=largest, sorted=sorted)
+        indices = torch.gather(indices, dim, k_indices)
+
+        metadata = a_parsed[0 : len_shape_a + 5]
+        final_result = torch.cat((metadata, result.double().flatten(), indices.double().flatten()))
+
+        b_parsed.copy_(final_result)
+
+    MPI_TOPK = MPI.Op.Create(mpi_topk, commute=True)
